refactor(web): split FinykApp.tsx — extract Nav, hash router, login screen
