+++ apps/web/src/modules/finyk/FinykApp.tsx
@@ -1,12 +1,9 @@
-import { useState, useEffect, useRef, useCallback } from "react";
+import { useState, useEffect, useRef } from "react";
 import { useMonobank } from "./hooks/useMonobank";
 import { usePrivatbank } from "./hooks/usePrivatbank";
 import { useStorage } from "./hooks/useStorage";
 import { readRaw, writeRaw } from "./lib/finykStorage";
 import { FINYK_MANUAL_ONLY_KEY, enableFinykManualOnly } from "./lib/demoData";
-import { PAGES } from "./constants";
-import { Button } from "@shared/components/ui/Button";
-import { Input } from "@shared/components/ui/Input";
 import { ModuleBottomNav } from "@shared/components/ui/ModuleBottomNav";
 import {
   ModuleHeader,
@@ -21,168 +18,14 @@ import { Budgets } from "./pages/Budgets";
 import { Assets } from "./pages/Assets";
 import { Analytics } from "./pages/Analytics";
 import { ManualExpenseSheet } from "./components/ManualExpenseSheet";
+import { FinykLoginScreen } from "./components/FinykLoginScreen";
+import { NAV_ICONS, NAV_IDS, NAV_ITEMS } from "./components/finykNav";
+import { useHashRouter, useHashQueryParam } from "./hooks/useHashRouter";
 import { useUnifiedFinanceData } from "./hooks/useUnifiedFinanceData";
 import { useFinykPersonalization } from "./hooks/useFinykPersonalization";
 import { useMonoTokenMigration } from "./hooks/useMonoTokenMigration";
 import { useFlag } from "../../core/lib/featureFlags";
 import { consumePresetPrefill } from "../../core/onboarding/presetPrefill";
-
-const NAV_ICONS = {
-  overview: (
-    <svg
-      width="22"
-      height="22"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6" />
-    </svg>
-  ),
-  transactions: (
-    <svg
-      width="22"
-      height="22"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M9 5H7a2 2 0 00-2 2v12a2 2 0 002 2h10a2 2 0 002-2V7a2 2 0 00-2-2h-2" />
-      <rect x="9" y="3" width="6" height="4" rx="1" />
-      <line x1="9" y1="12" x2="15" y2="12" />
-      <line x1="9" y1="16" x2="13" y2="16" />
-    </svg>
-  ),
-  budgets: (
-    <svg
-      width="22"
-      height="22"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <rect x="3" y="4" width="18" height="18" rx="2" ry="2" />
-      <line x1="16" y1="2" x2="16" y2="6" />
-      <line x1="8" y1="2" x2="8" y2="6" />
-      <line x1="3" y1="10" x2="21" y2="10" />
-    </svg>
-  ),
-  assets: (
-    <svg
-      width="22"
-      height="22"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <rect x="2" y="5" width="20" height="14" rx="2" />
-      <line x1="2" y1="10" x2="22" y2="10" />
-    </svg>
-  ),
-  analytics: (
-    <svg
-      width="22"
-      height="22"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <line x1="18" y1="20" x2="18" y2="10" />
-      <line x1="12" y1="20" x2="12" y2="4" />
-      <line x1="6" y1="20" x2="6" y2="14" />
-    </svg>
-  ),
-  settings: (
-    <svg
-      width="22"
-      height="22"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <circle cx="12" cy="12" r="3" />
-      <path d="M19.4 15a1.65 1.65 0 00.33 1.82l.06.06a2 2 0 010 2.83 2 2 0 01-2.83 0l-.06-.06a1.65 1.65 0 00-1.82-.33 1.65 1.65 0 00-1 1.51V21a2 2 0 01-4 0v-.09A1.65 1.65 0 009 19.4a1.65 1.65 0 00-1.82.33l-.06.06a2 2 0 01-2.83-2.83l.06-.06A1.65 1.65 0 004.68 15a1.65 1.65 0 00-1.51-1H3a2 2 0 010-4h.09A1.65 1.65 0 004.6 9a1.65 1.65 0 00-.33-1.82l-.06-.06a2 2 0 012.83-2.83l.06.06A1.65 1.65 0 009 4.68a1.65 1.65 0 001-1.51V3a2 2 0 014 0v.09a1.65 1.65 0 001 1.51 1.65 1.65 0 001.82-.33l.06-.06a2 2 0 012.83 2.83l-.06.06A1.65 1.65 0 0019.4 9a1.65 1.65 0 001.51 1H21a2 2 0 010 4h-.09a1.65 1.65 0 00-1.51 1z" />
-    </svg>
-  ),
-};
-
-const NAV_ITEMS = [
-  { id: "overview", label: "Огляд" },
-  { id: "transactions", label: "Операції" },
-  { id: "budgets", label: "Планування" },
-  { id: "analytics", label: "Аналітика" },
-  { id: "assets", label: "Активи" },
-];
-const NAV_IDS = NAV_ITEMS.map((n) => n.id);
-
-const ALL_PAGE_IDS = PAGES.map((p) => p.id);
-
-function useHashRouter(
-  defaultPage = "overview",
-): [string, (p: string) => void] {
-  const getPage = useCallback(() => {
-    // Accept `#page`, `#page?query`, legacy `#/page`. Query string is
-    // handled separately by `useHashQueryParam` — strip it here so it
-    // doesn't break page-id matching.
-    let p =
-      window.location.hash.replace(/^#\/?/, "").split(/[/?]/)[0] || defaultPage;
-    if (p === "payments") p = "budgets";
-    return p;
-  }, [defaultPage]);
-  const [page, setPageState] = useState(getPage);
-  useEffect(() => {
-    const handler = () => setPageState(getPage());
-    window.addEventListener("hashchange", handler);
-    return () => window.removeEventListener("hashchange", handler);
-  }, [getPage]);
-  const navigate = (p: string) => {
-    window.location.hash = p;
-  };
-  return [ALL_PAGE_IDS.includes(page) ? page : defaultPage, navigate];
-}
-
-/**
- * Reads a single query param from the URL hash (`#budgets?cat=smoking`
- * → `useHashQueryParam("cat") === "smoking"`). Used by deep-links from
- * Hub insights so a "Відкрити" tap can scroll the Budgets page to the
- * exact limit the recommendation is about.
- */
-function useHashQueryParam(name: string): string | null {
-  const read = useCallback(() => {
-    const m = window.location.hash.match(/\?(.+)$/);
-    if (!m) return null;
-    try {
-      return new URLSearchParams(m[1]).get(name);
-    } catch {
-      return null;
-    }
-  }, [name]);
-  const [value, setValue] = useState<string | null>(read);
-  useEffect(() => {
-    const handler = () => setValue(read());
-    window.addEventListener("hashchange", handler);
-    return () => window.removeEventListener("hashchange", handler);
-  }, [read]);
-  return value;
-}
 
 const PRIVAT_ENABLED = false;
 
@@ -369,249 +212,25 @@ export default function App({
   // ── Login screen ──────────────────────────────────────────────────────
   if (!clientInfo && !manualOnly) {
     return (
-      <div className="min-h-dvh flex items-center justify-center p-5 bg-bg safe-area-pt-pb">
-        <div className="w-full max-w-sm">
-          <div className="text-center mb-8">
-            <div
-              className={cn(
-                "w-20 h-20 mx-auto rounded-3xl flex items-center justify-center mb-4",
-                "bg-gradient-to-br from-brand-100 to-brand-200",
-                "dark:from-brand-900/40 dark:to-brand-800/30",
-                "border border-brand-200/60 dark:border-brand-700/30",
-                "shadow-card",
-              )}
-            >
-              <svg
-                width="40"
-                height="40"
-                viewBox="0 0 24 24"
-                fill="none"
-                stroke="currentColor"
-                strokeWidth="1.5"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                className="text-brand-600 dark:text-brand-400"
-                aria-hidden
-              >
-                <rect x="3" y="8" width="18" height="12" rx="2" />
-                <path d="M7 8V6a2 2 0 0 1 2-2h6a2 2 0 0 1 2 2v2" />
-                <line x1="3" y1="12" x2="21" y2="12" />
-              </svg>
-            </div>
-            <h1 className="text-2xl font-bold text-text">ФІНІК</h1>
-            <p className="text-sm text-muted mt-1">
-              Персональний фінансовий менеджер
-            </p>
-          </div>
-
-          <div
-            className={cn(
-              "bg-panel/95 backdrop-blur-xl border rounded-3xl p-6 shadow-float",
-              "border-line dark:border-line",
-            )}
-          >
-            <label
-              className="text-sm text-muted mb-2 block"
-              htmlFor="finyk-mono-token"
-            >
-              API токен Monobank
-            </label>
-            <p className="text-xs text-subtle mb-2">
-              Mono → Налаштування → Інші → API
-            </p>
-            <div className="relative mt-1">
-              <Input
-                id="finyk-mono-token"
-                className="pr-20"
-                type={showToken ? "text" : "password"}
-                placeholder="Вставте токен Mono API"
-                value={tokenInput}
-                onChange={(e) => setTokenInput(e.target.value)}
-                onKeyDown={(e) =>
-                  e.key === "Enter" &&
-                  connect(tokenInput.trim(), false, rememberToken)
-                }
-                autoComplete="off"
-              />
-              <Button
-                type="button"
-                size="sm"
-                variant="ghost"
-                className="absolute right-10 top-1/2 -translate-y-1/2 h-8 w-8 p-0 border-0"
-                aria-label="Вставити з буфера обміну"
-                title="Вставити з буфера"
-                onClick={async () => {
-                  try {
-                    setTokenInput(
-                      (await navigator.clipboard.readText()).trim(),
-                    );
-                  } catch {
-                    toast.error("Не вдалось прочитати буфер обміну");
-                  }
-                }}
-              >
-                <svg
-                  width="16"
-                  height="16"
-                  viewBox="0 0 24 24"
-                  fill="none"
-                  stroke="currentColor"
-                  strokeWidth="2"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  aria-hidden
-                >
-                  <rect x="9" y="9" width="13" height="13" rx="2" />
-                  <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
-                </svg>
-              </Button>
-              <Button
-                type="button"
-                size="sm"
-                variant="ghost"
-                className="absolute right-2 top-1/2 -translate-y-1/2 h-8 w-8 p-0 border-0"
-                aria-label={showToken ? "Приховати токен" : "Показати токен"}
-                onClick={() => setShowToken((v) => !v)}
-              >
-                {showToken ? (
-                  <svg
-                    width="16"
-                    height="16"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    aria-hidden
-                  >
-                    <path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19m-6.72-1.07a3 3 0 1 1-4.24-4.24" />
-                    <line x1="1" y1="1" x2="23" y2="23" />
-                  </svg>
-                ) : (
-                  <svg
-                    width="16"
-                    height="16"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    aria-hidden
-                  >
-                    <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" />
-                    <circle cx="12" cy="12" r="3" />
-                  </svg>
-                )}
-              </Button>
-            </div>
-
-            {webhookEnabled ? (
-              <p className="text-xs text-subtle mt-2">
-                Токен відправляється на сервер і не зберігається у браузері.
-              </p>
-            ) : (
-              <label className="flex items-center gap-2.5 mt-3 cursor-pointer select-none">
-                <input
-                  type="checkbox"
-                  className="w-4 h-4 rounded accent-emerald-600 cursor-pointer"
-                  checked={rememberToken}
-                  onChange={(e) => setRememberToken(e.target.checked)}
-                />
-                <span className="text-sm text-muted">
-                  Запам{"'"}ятати токен на цьому пристрої
-                </span>
-              </label>
-            )}
-
-            {authError && (
-              <div className="mt-3 text-sm bg-warning/15 border border-warning/40 rounded-xl px-3 py-2.5 space-y-1">
-                <p className="font-semibold text-text">
-                  Токен потребує оновлення
-                </p>
-                <p className="text-xs text-muted">{authError}</p>
-                <p className="text-xs text-muted">
-                  Отримайте новий токен: Monobank → Налаштування → API
-                </p>
-              </div>
-            )}
-            {error && !authError && (
-              <p className="mt-3 text-sm text-danger bg-danger/10 rounded-xl px-3 py-2">
-                {error}
-              </p>
-            )}
-
-            <Button
-              type="button"
-              className={cn(
-                "mt-4 w-full h-12 min-h-[48px] text-base border-0",
-                "bg-gradient-to-r from-brand-600 to-brand-700",
-                "hover:from-brand-700 hover:to-brand-800",
-                "text-white font-semibold",
-                "shadow-md hover:shadow-glow",
-                "transition-[background-color,box-shadow,opacity,transform] duration-200",
-                "active:scale-[0.98]",
-              )}
-              onClick={() => {
-                enableFinykManualOnly();
-                setManualOnly(true);
-              }}
-            >
-              Почати без банку
-            </Button>
-            <p className="mt-2 text-center text-xs text-subtle">
-              Ручні витрати, бюджети та аналітика — без API-токена. Monobank
-              можна підключити пізніше.
-            </p>
-
-            {/* eslint-disable-next-line sergeant-design/no-eyebrow-drift --
-                "або через API" divider row — structurally a delimiter
-                between two bg-line spans, not a heading. */}
-            <div className="my-4 flex items-center gap-3 text-xs text-muted uppercase tracking-wider">
-              <span className="flex-1 h-px bg-line" />
-              або через API
-              <span className="flex-1 h-px bg-line" />
-            </div>
-            <Button
-              type="button"
-              variant="secondary"
-              className="w-full min-h-[48px]"
-              onClick={() => connect(tokenInput.trim(), false, rememberToken)}
-              disabled={connecting || !tokenInput.trim()}
-            >
-              {connecting ? "Підключення…" : "Підключити Monobank"}
-            </Button>
-            {typeof onBackToHub === "function" && (
-              <Button
-                type="button"
-                variant="ghost"
-                className="mt-1 w-full min-h-[44px]"
-                onClick={onBackToHub}
-              >
-                ← Назад до хабу
-              </Button>
-            )}
-            <p className="mt-4 text-center text-xs text-subtle flex items-center justify-center gap-1.5">
-              <svg
-                width="12"
-                height="12"
-                viewBox="0 0 24 24"
-                fill="none"
-                stroke="currentColor"
-                strokeWidth="2"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                aria-hidden
-              >
-                <rect x="3" y="11" width="18" height="11" rx="2" />
-                <path d="M7 11V7a5 5 0 0 1 10 0v4" />
-              </svg>
-              Токен зберігається лише у твоєму браузері
-            </p>
-          </div>
-        </div>
-      </div>
+      <FinykLoginScreen
+        tokenInput={tokenInput}
+        onTokenInputChange={setTokenInput}
+        showToken={showToken}
+        onToggleShowToken={() => setShowToken((v) => !v)}
+        rememberToken={rememberToken}
+        onRememberTokenChange={setRememberToken}
+        webhookEnabled={webhookEnabled}
+        authError={authError}
+        error={error}
+        connecting={connecting}
+        onConnect={() => connect(tokenInput.trim(), false, rememberToken)}
+        onContinueWithoutBank={() => {
+          enableFinykManualOnly();
+          setManualOnly(true);
+        }}
+        toast={toast}
+        onBackToHub={onBackToHub}
+      />
     );
   }
